prototypes/dana/prototype/apps/cli/requirements.txt
# CLI mic → local Whisper (faster-whisper) → stdout
faster-whisper>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
sounddevice>=0.5.0
